backend/api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sse-starlette
langgraph
deepagents
langchain-openai
httpx
asyncpg
langgraph-checkpoint-postgres
python-jose[cryptography]
passlib[bcrypt]
